refactor: 배틀 목록 API 구조 개선 및 타입 가드 강화 >> >> - getBattleList.ts를 개별 파일로 분리 >>   - getOpenBattles.ts, getClosedBattles.ts로 분리 >>   - types.ts로 타입 정의 분리 >> - 타입 가드에서 타입 단언 제거 (모순적인거 같아서 지웠습니다.)
