+++ frontend/src/pages/mainPage/api/getOpenBattles.ts
@@ -1,0 +1,15 @@
+import { type GetBattleListParams, type BattleListResponse, isBattleListResponse } from './types';
+
+export async function getOpenBattles({ offset, limit }: GetBattleListParams): Promise<BattleListResponse> {
+  const res = await fetch(`/api/battles/open?offset=${offset}&limit=${limit}`);
+
+  if (!res.ok) {
+    throw new Error('배틀 목록을 불러오지 못했습니다.');
+  }
+
+  const data = await res.json();
+  if (isBattleListResponse(data)) {
+    return data;
+  }
+  throw new Error('잘못된 배틀 목록 응답 형식입니다.');
+}
